+++ packages/polaris-icons-internal/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Enhancements
+
+- Added plain SVG files in the `/images` folder for consumers that can't consume React components
 
 ### New icons